app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-multipart
passlib[bcrypt]
PyJWT
PyPDF2
httpx
google-genai
python-dotenv
python-magic